Move a helper to `rustc_trait_selection` and let `rustc_borrowck` use it from there
`rustc_borrowck` uses `type_op_prove_predicate_with_cause` from
`rustc_traits`, but is a simple helper that can live in
`rustc_trait_selection` instead. Move it there.
