+++ compiler/rustc_borrowck/src/diagnostics/bound_region_errors.rs
@@ -20,7 +20,8 @@ use rustc_span::Span;
 use rustc_trait_selection::error_reporting::InferCtxtErrorExt;
 use rustc_trait_selection::error_reporting::infer::nice_region_error::NiceRegionError;
 use rustc_trait_selection::traits::ObligationCtxt;
-use rustc_traits::{type_op_ascribe_user_type_with_span, type_op_prove_predicate_with_cause};
+use rustc_trait_selection::traits::query::type_op::prove_predicate::type_op_prove_predicate_with_cause;
+use rustc_traits::type_op_ascribe_user_type_with_span;
 use tracing::{debug, instrument};
 
 use crate::MirBorrowckCtxt;
