+++ compiler/rustc_trait_selection/src/traits/query/type_op/prove_predicate.rs
@@ -50,3 +50,15 @@ impl<'tcx> super::QueryTypeOp<'tcx> for ProvePredicate<'tcx> {
         Ok(())
     }
 }
+
+/// The core of the `type_op_prove_predicate` query: for diagnostics purposes in NLL HRTB errors,
+/// this query can be re-run to better track the span of the obligation cause, and improve the error
+/// message. Do not call directly unless you're in that very specific context.
+pub fn type_op_prove_predicate_with_cause<'tcx>(
+    ocx: &ObligationCtxt<'_, 'tcx>,
+    key: ParamEnvAnd<'tcx, ProvePredicate<'tcx>>,
+    cause: ObligationCause<'tcx>,
+) {
+    let ParamEnvAnd { param_env, value: ProvePredicate { predicate } } = key;
+    ocx.register_obligation(Obligation::new(ocx.infcx.tcx, cause, param_env, predicate));
+}
