+++ compiler/rustc_traits/src/lib.rs
@@ -15,7 +15,6 @@ mod type_op;
 
 use rustc_middle::query::Providers;
 pub use rustc_trait_selection::traits::query::type_op::ascribe_user_type::type_op_ascribe_user_type_with_span;
-pub use type_op::type_op_prove_predicate_with_cause;
 
 pub fn provide(p: &mut Providers) {
     dropck_outlives::provide(p);
